Fix landingPageImage
