+++ src/components/units/screens/board/write/boardWrite.container.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import BoardWriteUI from './boardWirte.presenter'
+import BoardWriteUI from './boardWrite.presenter'
 const BoardWrite = () => {
     return(
         <BoardWriteUI/>
